Rename util package to "internal" and fix build.
