+++ binder/src/main/java/io/grpc/binder/internal/Bindable.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.binder.util;
+package io.grpc.binder.internal;
 
 import android.os.IBinder;
 import androidx.annotation.AnyThread;
